// web/src/types/PluginCatalog.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PluginDetail } from "./PluginDetail";
import type { PluginError } from "./PluginError";

export type PluginCatalog = { plugins: Array<PluginDetail>, errors: Array<PluginError>, };