// web/src/types/PluginDetail.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PluginDescribe } from "./PluginDescribe";
import type { ProtocolFrame } from "./ProtocolFrame";

export type PluginDetail = { name: string, version: string | null, source: string, kind: string, binary: string, port: string, protocol_version: number, describe: PluginDescribe, transcript: Array<ProtocolFrame>, };